requirements.txt
streamlit==1.31.0
anthropic==0.25.0
python-dotenv==1.0.0
chromadb==0.4.22
langchain==0.1.10
langchain-community==0.0.24
sentence-transformers==2.5.1
pydantic==2.6.1
pydantic-settings==2.1.0
jieba==0.42.1
rank-bm25==0.2.2
